move store and related services access support code to its own store package, try to keep relatively clean separation of concerns, store should not know about local system decisions/layouts
