+++ daemon/api.go
@@ -47,6 +47,7 @@ import (
 	"github.com/ubuntu-core/snappy/release"
 	"github.com/ubuntu-core/snappy/snap"
 	"github.com/ubuntu-core/snappy/snappy"
+	"github.com/ubuntu-core/snappy/store"
 )
 
 // increase this every time you make a minor (backwards-compatible)
@@ -182,12 +183,12 @@ func sysInfo(c *Command, r *http.Request) Response {
 }
 
 type metarepo interface {
-	Snap(string, string) (*snappy.RemoteSnap, error)
-	FindSnaps(string, string) ([]*snappy.RemoteSnap, error)
+	Snap(string, string) (*store.RemoteSnap, error)
+	FindSnaps(string, string) ([]*store.RemoteSnap, error)
 }
 
 var newRemoteRepo = func() metarepo {
-	return snappy.NewUbuntuStoreSnapRepository()
+	return snappy.NewConfiguredUbuntuStoreSnapRepository()
 }
 
 var muxVars = mux.Vars
@@ -293,7 +294,7 @@ func getSnapsInfo(c *Command, r *http.Request) Response {
 	}
 
 	var localSnapsMap map[string][]*snappy.Snap
-	var remoteSnapMap map[string]*snappy.RemoteSnap
+	var remoteSnapMap map[string]*store.RemoteSnap
 
 	if includeLocal {
 		sources = append(sources, "local")
@@ -301,7 +302,7 @@ func getSnapsInfo(c *Command, r *http.Request) Response {
 	}
 
 	if includeStore {
-		remoteSnapMap = make(map[string]*snappy.RemoteSnap)
+		remoteSnapMap = make(map[string]*store.RemoteSnap)
 
 		// repo.Find("") finds all
 		//
